+++ lcurvetools/__init__.py
@@ -2,4 +2,4 @@ from .lcurves import lcurves, lcurves_by_MLP_estimator, lcurves_by_history
 from .utils import history_concatenate
 
 # Unique source of truth for the version number.
-__version__ = "1.1.0"
+__version__ = "1.1.1"